task-planner-ai-flow/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
google-auth>=2.16.2
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.86.0
python-multipart>=0.0.6